+++ .gitignore
@@ -12,4 +12,5 @@
 /captures
 .externalNativeBuild
 .cxx
+.idea
 local.properties
