+++ app/src/main/java/com/example/mustmarket/core/util/Resource.kt
@@ -1,4 +1,4 @@
-package com.example.mustmarket.util
+package com.example.mustmarket.core.util
 
 sealed class Resource<T>(val data: T? = null, val message: String? = null) {
     class Loading<T>(data: T? = null) : Resource<T>(data)
