+++ app/src/main/java/com/example/mustmarket/features/account/presentation/AccountScreen.kt
@@ -1,0 +1,9 @@
+package com.example.mustmarket.features.account.presentation
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+@Composable
+fun AccountScreen(modifier: Modifier = Modifier) {
+    
+}
