+++ app/src/main/java/com/example/mustmarket/features/auth/data/repository/AuthRepositoryImpl.kt
@@ -1,0 +1,4 @@
+package com.example.mustmarket.features.auth.data.repository
+
+class AuthRepositoryImpl {
+}
