+++ app/src/main/java/com/example/mustmarket/features/auth/domain/model/User.kt
@@ -1,0 +1,4 @@
+package com.example.mustmarket.features.auth.domain.model
+
+class User {
+}
