+++ app/src/main/java/com/example/mustmarket/features/auth/domain/repository/AuthRepository.kt
@@ -1,0 +1,4 @@
+package com.example.mustmarket.features.auth.domain.repository
+
+interface AuthRepository {
+}
