+++ app/src/main/java/com/example/mustmarket/features/auth/presentation/signup/SignUpViewModel.kt
@@ -1,0 +1,4 @@
+package com.example.mustmarket.features.auth.presentation.signup
+
+class SignUpViewModel {
+}
