+++ app/src/main/java/com/example/mustmarket/features/auth/presentation/splash/SplashScreenViewModel.kt
@@ -1,0 +1,4 @@
+package com.example.mustmarket.features.auth.presentation.splash
+
+class SplashScreenViewModel {
+}
