+++ app/src/main/java/com/example/mustmarket/features/cart/presentation/CartScreen.kt
@@ -1,0 +1,9 @@
+package com.example.mustmarket.features.cart.presentation
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+@Composable
+fun CartScreen(modifier: Modifier = Modifier) {
+
+}
