--- app/src/main/java/com/example/mustmarket/features/feature_login/presentation/login/LoginScreenState.kt
@@ -1,2 +1,0 @@
-package com.example.mustmarket.features.feature_login.presentation.login
-
