--- app/src/main/java/com/example/mustmarket/features/feature_register/presentation/signup/SignUpViewModel.kt
@@ -1,4 +1,0 @@
-package com.example.mustmarket.features.feature_register.presentation.signup
-
-class SignUpViewModel {
-}
