--- app/src/main/java/com/example/mustmarket/features/feature_register/presentation/signup/SignUpViewModelState.kt
@@ -1,2 +1,0 @@
-package com.example.mustmarket.features.feature_register.presentation.signup
-
