--- app/src/main/java/com/example/mustmarket/features/feature_splash/SplashScreenViewModel.kt
@@ -1,4 +1,0 @@
-package com.example.mustmarket.features.feature_splash
-
-class SplashScreenViewModel {
-}
