+++ app/src/main/java/com/example/mustmarket/navigation/NavGraph.kt
@@ -10,9 +10,9 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import coil.annotation.ExperimentalCoilApi
-import com.example.mustmarket.features.feature_login.presentation.login.LoginScreen
-import com.example.mustmarket.features.feature_register.presentation.signup.SignUpScreen
-import com.example.mustmarket.features.feature_splash.SplashScreen
+import com.example.mustmarket.features.auth.presentation.login.LoginScreen
+import com.example.mustmarket.features.auth.presentation.signup.SignUpScreen
+import com.example.mustmarket.features.auth.presentation.splash.SplashScreen
 
 @ExperimentalCoilApi
 @ExperimentalFoundationApi
@@ -21,15 +21,13 @@ import com.example.mustmarket.features.feature_splash.SplashScreen
 fun SetUpNavGraph(
     navController: NavHostController
 ) {
-    Scaffold { innerPadding ->
-        NavHost(
-            modifier = Modifier.padding(innerPadding),
-            navController = navController,
-            startDestination = Screen.Splash.route
-        ) {
-            composable(route = Screen.SignUp.route) { SignUpScreen(navController = navController) }
-            composable(route = Screen.Login.route) { LoginScreen(navController = navController) }
-            composable(route = Screen.Splash.route) { SplashScreen(navController = navController) }
-        }
+    // mbona unaeka pading kwa routing ;)
+    NavHost(
+        navController = navController,
+        startDestination = Screen.Splash.route
+    ) {
+        composable(route = Screen.SignUp.route) { SignUpScreen(navController = navController) }
+        composable(route = Screen.Login.route) { LoginScreen(navController = navController) }
+        composable(route = Screen.Splash.route) { SplashScreen(navController = navController) }
     }
 }
